requirements.txt
yamldataclassconfig==1.5.0
flask==3.0.0
pytest==7.4.4
yamldataclassconfig==1.5.0
openai==1.6.1
python-docx==1.1.0